+++ src/app/components/Accordion/AccordionBody/AccordionBody.tsx
@@ -1,0 +1,9 @@
+import { PropsWithChildren } from "react";
+
+import { useAccordionContext } from "../AccordionContext";
+
+export default function AccordionBody({ children }: PropsWithChildren) {
+  const { isOpen } = useAccordionContext();
+
+  return isOpen ? <div>{children}</div> : null;
+}
