--- src/app/workspace/[id]/components/ui/AccordionPanel/AccordionPanelBody/AccordionPanelBody.tsx
@@ -1,9 +1,0 @@
-import { PropsWithChildren } from "react";
-
-import { useAccordionPanelContext } from "../AccordionPanelContext";
-
-export default function AccordionPanelBody({ children }: PropsWithChildren) {
-  const { isOpen } = useAccordionPanelContext();
-
-  return isOpen ? <div>{children}</div> : null;
-}
