+++ src/app/workspace/[id]/components/ui/PanelContainer/PanelContainer.tsx
@@ -1,0 +1,11 @@
+import { PropsWithChildren } from "react";
+
+import styles from "./PanelContainerStyle.module.scss";
+import classNames from "classnames/bind";
+
+const cx = classNames.bind(styles);
+
+export default function PanelContainer({ children }: PropsWithChildren) {
+  return <div className={cx("container")}>{children}</div>;
+}
+
