+++ src/app/workspace/[id]/components/ui/PanelContainer/PanelContainerStyle.module.scss
@@ -2,7 +2,4 @@
   padding: 16px;
   border-radius: 16px;
   background: #f9fbfd;
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
 }
